moved fem to be a sub-module within amigo
