+++ amigo/fem/__init__.py
@@ -1,0 +1,15 @@
+from .fem import Problem, Mesh
+from .basis import SolutionSpace, dot_product, curl_2d, mat_vec, mat_vec_transpose
+from .connectivity import InpParser, plot_mesh
+
+__all__ = [
+    Problem,
+    Mesh,
+    SolutionSpace,
+    InpParser,
+    plot_mesh,
+    dot_product,
+    curl_2d,
+    mat_vec,
+    mat_vec_transpose,
+]
